+++ package.json
@@ -35,6 +35,7 @@
     "@types/pino": "^5.14.0",
     "@types/sinon": "^10.0.2",
     "aws-sdk-client-mock": "^3.0.0",
+    "aws-sdk-client-mock-jest": "^3.0.0",
     "graceful-fs": "^4.2.3",
     "husky": "^1.3.1",
     "jest": "^29.6.2",
